feat: Add `fisher_yates_shuffle` method and refactor workspace
* move no-std rand functions to `no_std_rand.rs`
* move AFS types to `afs.rs`
* Add `fisher_yates_shuffle` function that shuffles list elements
